feat: implement basic poll mechanism and refined process blocking semantics
Summary of Changes:

1. Standardized File I/O & Flag Infrastructure:
   - Upgraded `flags` field to `int32_t` across `sys_open`, `file_open`, and
     `struct file` to accommodate extended POSIX open modes.
   - Introduced `O_NONBLOCK` (128) support, enabling non-blocking I/O
     capabilities for character devices TTY.
   - Implemented a bidirectional flag translation layer in `do_fcntl` and
     `do_open` to bridge internal MagicBox constants with standard
     Linux/musl-libc definitions.

2. Refined Process Blocking Semantics:
   - Formalized distinct process states: `TASK_BLOCKED` for uninterruptible
     kernel waits and `TASK_WAITING` for interruptible blocking (signals).
   - Refactored `sys_milsleep` and `sys_pause` to utilize `TASK_WAITING`,
     allowing proper signal delivery during synchronous delays.

3. Concurrency Control & Semaphore Enhancements:
   - Introduced `sema_try_wait`, a non-blocking semaphore acquisition
     primitive that returns immediate status instead of yielding the CPU.
   - Optimized `tty_read` to utilize `sema_try_wait`, preventing race
     conditions and unexpected blocking when multiple processes (e.g.,
     blocking vs non-blocking readers) contend for the TTY, thereby
     preserving strict `poll` semantics.

4. Implementation of POSIX-compliant Poll Mechanism:
   - Developed the core `poll` subsystem in `fs/poll.c`, replacing the
     legacy "fake" interceptor with a true state-aware implementation.
   - Leveraged `poll_table` and `poll_wait` hooks to allow processes to
     efficiently sleep on multiple device wait-queues.
   - Successfully enabled fully functional `vi` and interactive shell
     environments; `ash` now correctly manages TTY state and handles
     asynchronous user input via a stable syscall loop.

5. Validation & Testing:
   - Added `prog/musl_test/test_fcntl.c` to verify the accuracy of the
     syscall flag translation and the robustness of the `F_SETFL` logic.
   - Fixed a critical edge case where `timeout = -1` (infinite wait) in
     `poll` incorrectly triggered immediate returns due to 32-bit integer
     overflow in the sleep timer calculation.

+++ fs/file.c
@@ -62,7 +62,7 @@ int32_t file_close(struct file* file){
 	return 0;
 }
 
-int32_t file_open(struct partition* part, uint32_t inode_no,uint8_t flag){
+int32_t file_open(struct partition* part, uint32_t inode_no,int32_t flag){
 	int32_t fd_idx = get_free_slot_in_global();
 	if(fd_idx==-1){
 		printk("exceed max open files\n");
